+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+#
+* Deserialize Vec and HashMap to Option in case a value is null and unwrap a default if it is missing
+
 # 0.7.0
 * Make futures returned by methods Send + Sync